requirements.txt
gitpython
loguru
tiktoken
pyyaml
scipy
playsound
websockets
pyaudio